+++ src/tracker/constants.ts
@@ -9,6 +9,6 @@ export const IMMEDIATE_PREVIOUS_APP_VERSION = "0.2.0";
 
 export const DEFAULT_SETTINGS: TrackerSettings = {
   showMinimap: true,
-  defaultArrowMode: "forward",
+  defaultArrowMode: "bidirectional",
   hidePlacedLocations: false,
 };